[#31] 보낸 제안서 확인 UI 구현
- 페이지 이동 시 Tab bar 숨기기 구현
(but, 다시 홈으로 돌아갔을 때 로딩 시간 고려해 봐야 함)
- 홈 버튼, 프로필 연결, 공유하기 미완성
